plus que les builtin et relire le replace

+++ srcs/libft/ft_strndup.c
@@ -6,7 +6,7 @@
 /*   By: ntom <marvin@42.fr>                        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/09/06 13:58:40 by ntom              #+#    #+#             */
-/*   Updated: 2018/09/06 14:02:30 by ntom             ###   ########.fr       */
+/*   Updated: 2019/05/24 20:21:59 by ntom             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strndup(const char *s1, size_t n)
 {
 	char			*str;
-	int				i;
+	size_t			i;
 
 	i = 0;
 	if (!(str = (char*)malloc(sizeof(char) * (n + 1))))
